fix(core): create new env variable for performance profile and add documentation
